rename : 클래스 명, 디렉토리 구조 request, response 추가
