+++ src/main/java/ddd/caffeine/ratrip/module/place/presentation/dto/bookmark/BookmarkPlaceResponseDto.java
@@ -8,7 +8,7 @@ import lombok.Getter;
 
 @Getter
 public class BookmarkPlaceResponseDto {
-	private List<BookmarkPlaceModel> places;
+	private List<BookmarkPlaceResponse> places;
 	private boolean hasNext;
 
 	public BookmarkPlaceResponseDto(List<BookMarkPlaceDao> places, boolean hasNext) {
@@ -19,7 +19,7 @@ public class BookmarkPlaceResponseDto {
 	private void mapBookMarkResponseModels(List<BookMarkPlaceDao> places) {
 		this.places = new ArrayList<>();
 		for (BookMarkPlaceDao place : places) {
-			this.places.add(new BookmarkPlaceModel(place));
+			this.places.add(new BookmarkPlaceResponse(place));
 		}
 	}
 }
