+++ src/main/java/ddd/caffeine/ratrip/module/place/presentation/dto/request/PlaceSaveByThirdPartyRequestDto.java
@@ -1,4 +1,4 @@
-package ddd.caffeine.ratrip.module.place.presentation.dto;
+package ddd.caffeine.ratrip.module.place.presentation.dto.request;
 
 import javax.validation.constraints.NotBlank;
