+++ src/main/java/ddd/caffeine/ratrip/module/place/presentation/dto/response/PlaceDetailResponseDto.java
@@ -1,12 +1,11 @@
-package ddd.caffeine.ratrip.module.place.presentation.dto;
+package ddd.caffeine.ratrip.module.place.presentation.dto.response;
 
 import java.util.List;
 import java.util.UUID;
 
+import ddd.caffeine.ratrip.module.place.domain.Place;
 import ddd.caffeine.ratrip.module.place.domain.sub_domain.Blog;
 import ddd.caffeine.ratrip.module.place.domain.sub_domain.Location;
-import ddd.caffeine.ratrip.module.place.domain.Place;
-import ddd.caffeine.ratrip.module.place.presentation.dto.bookmark.BookmarkResponseDto;
 import lombok.Getter;
 
 /**
