+++ src/main/java/ddd/caffeine/ratrip/module/place/presentation/dto/response/PlaceInRegionResponse.java
@@ -1,9 +1,9 @@
-package ddd.caffeine.ratrip.module.place.presentation.dto;
+package ddd.caffeine.ratrip.module.place.presentation.dto.response;
 
 import java.util.UUID;
 
-import ddd.caffeine.ratrip.module.place.domain.sub_domain.Location;
 import ddd.caffeine.ratrip.module.place.domain.Place;
+import ddd.caffeine.ratrip.module.place.domain.sub_domain.Location;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -12,14 +12,14 @@ import lombok.NoArgsConstructor;
  */
 @Getter
 @NoArgsConstructor
-public class PlaceInRegionModel {
+public class PlaceInRegionResponse {
 	private UUID id;
 	private String name;
 	private String category;
 	private String imageLink;
 	private Location location;
 
-	public PlaceInRegionModel(Place place) {
+	public PlaceInRegionResponse(Place place) {
 		this.id = place.getId();
 		this.name = place.getName();
 		this.category = place.getCategory().name();
