+++ src/main/java/ddd/caffeine/ratrip/module/place/presentation/dto/response/PlaceInRegionResponseDto.java
@@ -1,4 +1,4 @@
-package ddd.caffeine.ratrip.module.place.presentation.dto;
+package ddd.caffeine.ratrip.module.place.presentation.dto.response;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -10,7 +10,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @NoArgsConstructor
 public class PlaceInRegionResponseDto {
-	private List<PlaceInRegionModel> placeInRegionModels;
+	private List<PlaceInRegionResponse> placeInRegionResponses;
 	private boolean hasNext;
 
 	public PlaceInRegionResponseDto(List<Place> places, boolean hasNext) {
@@ -19,9 +19,9 @@ public class PlaceInRegionResponseDto {
 	}
 
 	private void create(List<Place> places) {
-		this.placeInRegionModels = new ArrayList<>();
+		this.placeInRegionResponses = new ArrayList<>();
 		for (Place place : places) {
-			placeInRegionModels.add(new PlaceInRegionModel(place));
+			placeInRegionResponses.add(new PlaceInRegionResponse(place));
 		}
 	}
 }
